Type BasketHeader, CasinoEffect, UserContext, helper
