+++ src/fakeData/fakeMenu.tsx
@@ -1,12 +1,4 @@
-export type MenuType = {
-  id: string,
-  imageSource: string,
-  title: string,
-  price: number,
-  quantity: number,
-  isAvailable: boolean,
-  isPublicised: boolean,
-}
+import { MenuType } from "../typeScript/MenuType";
 
 const EMPTY: MenuType[] = [];
