+++ src/helper/helper.tsx
@@ -1,8 +1,9 @@
+import { ProductType } from "../enums/product";
 import { findObjectById } from "../utils/array";
 import { convertStringToBoolean } from "../utils/string";
 
-export const calculateBasketTotal = (menu, basketProduct) => {
-  const total = basketProduct.reduce((total, basketProduct) => {
+export const calculateBasketTotal = (menu: ProductType[], basketProduct: ProductType[]): number => {
+  const total = basketProduct.reduce((total: number, basketProduct: ProductType) => {
     const menuProduct = findObjectById(menu, basketProduct.id);
     if (isNaN(menuProduct.price)) return total;
     if (convertStringToBoolean(menuProduct.isAvailable) === false) return total;
@@ -12,5 +13,5 @@ export const calculateBasketTotal = (menu, basketProduct) => {
   return total;
 };
 
-export const checkProductIsClicked = (idProductedCliked, array) =>
+export const checkProductIsClicked = (idProductedCliked: string, array: ProductType) =>
   idProductedCliked === array.id;
