+++ src/typeScript/MenuType.ts
@@ -1,0 +1,9 @@
+export type MenuType = {
+    id: string,
+    imageSource: string,
+    title: string,
+    price: number,
+    quantity: number,
+    isAvailable: boolean,
+    isPublicised: boolean,
+}
